Implement refinement strategies for waveFieldWriter
The refinement allows the mesh to be subdivided into further cells
before it is written to the output file. After the initialization
process, given the variables values at the vertices of a cell, their
values at the centers of the finer cell level can be extracted.

Refinement strategies in the current implementation :

 * 1 -> 1  Pass-through without any refinement.
 * 1 -> 4  Center-split results in 4 cells of equal volume.
 * 1 -> 8  Face-split results in 8 cells whereas the face areas will
           be split evenly. Results in 4 outer/corner and 4 inner cells.
 * 1 -> 32 Applies Face-split and center-split in succession.

Refinement can be turned on using the parameter file.
Valid values for the parameter are 0-3 using the listing above.
0 stands for 1->1 and 3 for 1->32 mapping.