Add surface writer to elastic-acoustic interface. Reorder dg init.
FreeSurfaceWriter now also writes the data along an
elastic-acoustic interface.
The intialization of the clustered LTS is now split into two parts
where the first fixates the LTS tree and the second inits the
buckets (and some other stuff).
The material parameters are now available in the second method and
can thus be used to detect the elastic-acoustic interface during
construction of the bucket.