Added a report of a GPU memory status before entering the compute-loop
* added a check for device memory. A user can allocate more memory than
available on a GPU e.g., unified or pinned memory. If a user allocates
lots of unified memory then it can lead to frequent page faults
which will kill the entire GPU performance. If it happens in SeisSol we
will abort such an execution to prevent burning GPU hours on an HPC cluster

* moved deep copying from host to device to a dedicated function