Compute face displacement via integrals for fs/e-a
Compute face displacements by projection of integrated
quantities to face instead of pointlessly solving ODE.

Store face displacements in global coordinate system.
As the grav. free surface bc computes it in
face-normal coordinate system, we need to rotate
the displacement twice in this kernel.
This is very cheap (rotation of vector of size
numb. of 2D basis x 3).